fixed float bug

+++ src/models.rs
@@ -4,12 +4,12 @@ use super::schema::expenses;
 pub struct Expense {
     pub id: i32,
     pub name: String,
-    pub amount: i32,
+    pub amount: f32,
 }
 
 #[derive(Insertable)]
 #[table_name = "expenses"]
 pub struct NewExpense<'a> {
     pub name: &'a str,
-    pub amount: i32,
+    pub amount: f32,
 }
