+++ src/schema.rs
@@ -2,6 +2,6 @@ table! {
     expenses (id) {
         id -> Integer,
         name -> Text,
-        amount -> Integer,
+        amount -> Float,
     }
 }
